M4 Wave 6 AMS1: STRICT_COMPILERFLAGS for ams/libs/ms, 3 genuine bugs fixed
FossilOrigin-Name: 61b4a32ebaf913b436b5a279ca393a2894e541760b1eaeaba38f673515a19772

+++ src/ams/libs/ms/logging.c
@@ -77,7 +77,7 @@ int ConsiderLoggingRead(char *FileName)
     fclose(fp);
     s = strrchr(FileName, '/');
     if (s) *s = '\0';
-    fprintf(logfp, "X-StatTrace: %s READ %d bytes %s %s ; %s", Me, statbuf.st_size - size, strncmp(FileName, home, strlen(home)) ? "BBOARD" : "MAIL", FileName, arpadate());
+    fprintf(logfp, "X-StatTrace: %s READ %ld bytes %s %s ; %s", Me, (long)(statbuf.st_size - size), strncmp(FileName, home, strlen(home)) ? "BBOARD" : "MAIL", FileName, arpadate());
     if (s) *s = '/';
     fclose(logfp);
 }
